[business logic] Move our msgspec models in "models.py"
Even though they're not Django models, semantically speaking they *are* models :-)
