+++ src/apps/daily_challenge/business_logic/_manage_daily_challenge_moved_piece_logic.py
@@ -2,10 +2,10 @@ from typing import TYPE_CHECKING
 
 from django.utils.timezone import now
 
-from ..types import PlayerGameOverState
+from ..models import PlayerGameOverState
 
 if TYPE_CHECKING:
-    from ..types import PlayerGameState, PlayerStats
+    from ..models import PlayerGameState, PlayerStats
 
 
 def manage_daily_challenge_moved_piece_logic(
